Add B+tree to enterprise subpackage
