+++ ctl/inspect.go
@@ -69,7 +69,7 @@ func (cmd *InspectCommand) Run(ctx context.Context) error {
 	// Attach the mmap file to the bitmap.
 	t := time.Now()
 	fmt.Fprintf(cmd.Stderr, "unmarshaling bitmap...")
-	bm := roaring.NewSliceBitmap()
+	bm := roaring.NewBitmap()
 	if err := bm.UnmarshalBinary(data); err != nil {
 		return errors.Wrap(err, "unmarshalling")
 	}
